Move a few RDG tests across to shared infrastructure.
